refactor(krites/fts): full structural overhaul — snafu errors, quality (#3217)
- Split mod.rs: extract TokenizerConfig construction into config.rs (280 LOC)
- Add InvalidNgramConfig variant to FtsError for proper n-gram validation
- Replace assert!() panics in NgramTokenizer with debug_assert (callers validate)
- Add Debug derive to all tokenizer/filter types (Token, AlphaNumOnlyFilter,
  LowerCaser, RawTokenizer, SimpleTokenizer, WhitespaceTokenizer, NgramTokenizer,
  RemoveLongFilter, Stemmer, AsciiFoldingFilter, EmptyTokenizer)
- Add Hash derive to Token and Language for use in collections
- Document BM25 scoring formula inline with full mathematical notation
- Document TF/TF-IDF scoring function
- Add doc comments to all tokenizer stages, trait definitions, and wrapper types
- Remove stale tantivy references from doc comments throughout
- Fix typo: token_steam -> token_stream in TextAnalyzer::unique_ngrams
- Add doc comments to FtsIndexManifest, TokenizerConfig, TokenizerCache,
  SplitCompoundWords, StopWordFilter
